requirements.txt
alembic==1.4.3
better-profanity==0.6.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docopt==0.6.2
Flask==1.1.2
Flask-Script==2.0.6
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
password-validator==1.0
pluggy==0.13.1
pyparsing==2.4.7
pytest==6.0.2
pytest-flask==1.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
yarg==0.1.9
zipp==3.1.0
